+++ src/controles/control_compra.java
@@ -19,6 +19,7 @@ import clases.inventario;
 import conexion.conexion;
 import consultas.consultas_compra;
 import consultas.consultas_inventario;
+import formularios.login_usuario;
 import formularios.registro_compras;
 import formularios.registro_inventario;
 
@@ -269,21 +270,27 @@ public class control_compra implements ActionListener {
 				if (filaseleccionada == -1) {
 					JOptionPane.showMessageDialog(null, "No se ha seleccionado ninguna fila");
 				} else {
-					conexion objCon = new conexion();
-					Connection conn = objCon.getConexion();
-					int Fila = formulario.tabla.getSelectedRow();
-					String codigo = formulario.tabla.getValueAt(Fila, 0).toString();
-					ps = conn.prepareStatement("DELETE FROM compras WHERE id_compra=?");
-					ps.setString(1, codigo);
-					ps.execute();
-					JOptionPane.showMessageDialog(null, "Objeto Eliminado!");
-					limpiar();
-					formulario.construirTabla();
-					formulario.txtCodigo.setText(null);
-					formulario.btnAceptar.setEnabled(true);
-					formulario.btnActualizar.setVisible(false);
-					formulario.btnGuardar.setVisible(false);
-					formulario.btnNuevo.setVisible(false);
+					if (login_usuario.cargoUsuario.toString() == "Usuario Avanzado") {
+
+						conexion objCon = new conexion();
+						Connection conn = objCon.getConexion();
+						int Fila = formulario.tabla.getSelectedRow();
+						String codigo = formulario.tabla.getValueAt(Fila, 0).toString();
+						ps = conn.prepareStatement("DELETE FROM compras WHERE id_compra=?");
+						ps.setString(1, codigo);
+						ps.execute();
+						JOptionPane.showMessageDialog(null, "Objeto Eliminado!");
+						limpiar();
+						formulario.construirTabla();
+						formulario.txtCodigo.setText(null);
+						formulario.btnAceptar.setEnabled(true);
+						formulario.btnActualizar.setVisible(false);
+						formulario.btnGuardar.setVisible(false);
+						formulario.btnNuevo.setVisible(false);
+					} else {
+						JOptionPane.showMessageDialog(null,
+								"Usted no tiene permisos para eliminar (Solo el jefe de la empresa)");
+					}
 
 				}
 			} catch (SQLException ex) {
